Better error handling
In rare cases, the *_MASTER tables might include entries for
records that don't actually exist and have no MARC in the
*_DATA tables. This was causing a StopIteration in the methods that
read records from the database when next() was used to read from
a pymarc.MARCReader, because there was no record to read. In turn,
this would cause for loops in client code to silently stop running
before they reached the end of the records they should have been
reading.

Now an exception will be raised in get_bib and get_mfhd if no
data is returned.